Modify: combine the crates
