--- core/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "lt-fm-index-core"
-version = "0.1.0"
-authors = ["baku4 <bahkhun@gmail.com>"]
-edition = "2021"
-rust-version = "1.57.0"
-
-[dependencies]
-thiserror = "1.0.38"
-byteorder = "1.4.3"
-num-traits = "0.2.15"
-capwriter = "0.1.4"
-bytemuck = { version = "1.13.0", features = [ "derive" ] }
-rand = "0.8.0"
-bio = "1.1.0"
-libdivsufsort-rs = { version = "0.1.2", optional = true }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-getrandom = { version = "0.2", features = ["js"] }
-
-[dev-dependencies]
-
-[features]
-default = []
-fastbwt = ["libdivsufsort-rs"]
