--- core/src/lib.rs
@@ -1,14 +1,0 @@
-/*!
-Library for core components of lt-fm-index
-*/
-mod core;
-
-mod algorithm;
-// pub use algorithm::{
-//     RawLtFmIndex,
-//     Block,
-// };
-// // pub mod blocks;
-
-// #[cfg(test)]
-// mod tests;
