--- lt-fm-index/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "lt-fm-index"
-version = "0.6.0"
-authors = ["baku4 <bahkhun@gmail.com>"]
-edition = "2021"
-rust-version = "1.57.0"
-description = "Fm-index using k-mer lookup table for exact pattern matching"
-readme = "README.md"
-repository = "https://github.com/baku4/lt-fm-index/"
-license = "MIT"
-keywords = ["fm-index", "bwt", "pattern_matching", "bio"]
-categories = ["science", "algorithms"]
-
-[dependencies]
-lt-fm-index-bridge-32 = { path = "../bridges/v32" }
-lt-fm-index-bridge-64 = { path = "../bridges/v64" }
