--- lt-fm-index/src/lib.rs
@@ -1,9 +1,0 @@
-use lt_fm_index_bridge_32::*;
-use lt_fm_index_bridge_64::*;
-
-#[test]
-fn test() {
-    // let v32 = RawLtFmIndex32::<T32V2u32>::new(
-        
-    // );
-}
