+++ src/algorithm/bwm/blocks/mod.rs
@@ -1,0 +1,15 @@
+use super::Block;
+
+mod vector;
+mod block2;
+mod block3;
+mod block4;
+mod block5;
+mod block6;
+
+pub use vector::Vector;
+pub use block2::Block2;
+pub use block3::Block3;
+pub use block4::Block4;
+pub use block5::Block5;
+pub use block6::Block6;
