+++ src/algorithm/suffix_array/burrow_wheeler_transform/libdivsufsort.rs
