+++ src/algorithm/suffix_array/burrow_wheeler_transform/mod.rs
