+++ src/algorithm/suffix_array/bwt.rs
