+++ src/algorithm/suffix_array/count_array.rs
