+++ src/wrapper/mod.rs
@@ -1,0 +1,1 @@
+use super::{LtFmIndex, blocks::*};
